\name{PERICLIMv1.0-package}

\alias{PERICLIMv1.0-package}

\title{PERICLIMv1.0}

\description{
PERICLIMv1.0 is a simple modelling scheme that allows to derive palaeo-air temperature characteristics associated with relict periglacial features indicative of the palaeo-active-layer thickness.
}

\details{
The package consists of two alternate schemes deriving past air temperature conditions that differ in how they define the magnitude of annual air temperature oscillations, but both provide identical outcomes if driven with compatible data.

\code{\link{Aa}} utilizes the temperature range (Aa) to define the magnitude of annual air temperature oscillations.

\code{\link{MATWM}} utilizes the mean air temperature of the warmest month (MATWM) to define the magnitude of annual air temperature oscillations.
}
